refactor: move polyline6 decoder to utils
